디퍼드 렌더러로 변경
todo
그러나 materialIdx가 오작동 하는 것을 발견
자고 일어나서 고치자

+++ BLAH/BlahBlah/COOL/Include/Mesh/Mesh.cpp
@@ -197,7 +197,7 @@ void Mesh::Render(ComPtr<ID3D12GraphicsCommandList> commandList, XMFLOAT4X4& par
 
 		XMFLOAT4X4 temp = Matrix4x4::Transpose(m_RootTransform);
 
-		commandList->SetGraphicsRoot32BitConstants(ROOT_SIGNATURE_IDX::WORLD_MATRIX, 16, &temp, 0);
+		commandList->SetGraphicsRoot32BitConstants(static_cast<int>(ROOT_SIGNATURE_IDX::WORLD_MATRIX), 16, &temp, 0);
 
 		commandList->IASetVertexBuffers(0, _countof(vertexBufferViews), vertexBufferViews);
